🚚  Move ROS-related files to ros/ folder
