+++ src/lib/data/test/robot_state.test.cpp
@@ -1,4 +1,13 @@
-// Bring in my package's API, which is what I'm testing
+/**
+ * @file robot_state.test.cpp
+ * @author Felipe Gomes de Melo <felipe.gomes@thunderatz.org>
+ * @brief
+ * @date 04/2021
+ *
+ * @copyright MIT License - Copyright (c) 2021 ThundeRatz
+ *
+ */
+
 #include <ros/ros.h>
 #include <ros/console.h>
 #include <gtest/gtest.h>
@@ -7,7 +16,7 @@
 #include <iostream>
 
 #include "travesim_adapters/data/robot_state.hpp"
-#include "travesim_adapters/vision_receiver.hpp"
+#include "travesim_adapters/ros/vision_receiver.hpp"
 
 // Declare a test
 TEST(robot_state_to_model_state, model_name)
